refactor(vscode): introduce `ToolInterface` (#15974)
> This PR refactors the VSCode extension's linter functionality by introducing a ToolInterface abstraction. The refactoring moves from function-based exports to a class-based architecture, preparing the codebase for supporting multiple tools (e.g., linter and formatter) with a consistent interface.
